+++ how-to/workspace-platform-starter/client/src/framework/platform/interop.ts
@@ -6,7 +6,7 @@ import type { PlatformInteropOverride, PlatformInteropOverrideOptions } from "..
 import type { ModuleDefinition, ModuleEntry, ModuleHelpers } from "../shapes/module-shapes";
 import type { PlatformProviderOptions } from "../shapes/platform-shapes";
 import { isEmpty } from "../utils";
-import { getConstructorOverride } from "./broker/interopbroker";
+import { getConstructorOverride } from "./broker/interop-override";
 
 const logger = createLogger("InteropProvider");
 const allOverrides: OpenFin.ConstructorOverride<OpenFin.InteropBroker>[] = [];
@@ -25,32 +25,49 @@ export async function init(
 	helpers: ModuleHelpers
 ): Promise<void> {
 	if (options) {
-		const interopOverrideSettings: PlatformInteropOverrideOptions | undefined = options?.interop;
-		if(!isEmpty(interopOverrideSettings)) {
-			interopOverrideSettings.windowPositionOptions = windowPositioningOptions;
-			interopOverrideSettings.platformRootUrl = options?.rootUrl;
-		}
-		if(Array.isArray(options?.interop?.modules)) {
+		// eslint-disable-next-line @typescript-eslint/no-unused-vars
+		const { modules: moduleOptions, ...settings } = options?.interop ?? {};
+		const interopOverrideSettings: PlatformInteropOverrideOptions | undefined = settings;
+
+		interopOverrideSettings.windowPositionOptions = windowPositioningOptions;
+		interopOverrideSettings.platformRootUrl = options?.rootUrl;
+		interopOverrideSettings.intentOptions = {
+			intentTimeout: 15000,
+			...options?.interop?.intentOptions
+		};
+		interopOverrideSettings.intentResolver = interopOverrideSettings?.intentResolver ?? {
+			fdc3InteropApi: "2.0",
+			url: `${options.rootUrl}/common/windows/intents/instance-picker.html`,
+			title: "Intent Resolver"
+		};
+		interopOverrideSettings.openOptions = {
+			openStrategy: "fdc3",
+			...interopOverrideSettings?.openOptions
+		};
+
+		if (Array.isArray(moduleOptions)) {
 			modules = await loadModules<PlatformInteropOverride>(options.interop, "interopOverride");
 			await initializeModules<PlatformInteropOverride>(modules, helpers);
 		}
 
 		logger.info("Getting interop overrides...");
-		// eslint-disable-next-line @typescript-eslint/no-unused-vars
-		const { modules: moduleOptions, ...settings } = options?.interop ?? {};
 
-		if(settings?.defaultBrokerStrategy === "first") {
+		if (
+			interopOverrideSettings?.defaultBrokerStrategy === "first" ||
+			isEmpty(interopOverrideSettings?.defaultBrokerStrategy)
+		) {
 			logger.info("Adding default interop override first");
-			allOverrides.push(await getConstructorOverride(settings));
+			allOverrides.push(await getConstructorOverride(interopOverrideSettings));
 		}
 		for (const interopModule of modules) {
-			const interopConstructor = await interopModule.implementation.getConstructorOverride(settings);
+			const interopConstructor =
+				await interopModule.implementation.getConstructorOverride(interopOverrideSettings);
 			allOverrides.push(interopConstructor);
 			logger.info(`Added interopOverride module: ${interopModule.definition.id}`);
 		}
-		if(settings?.defaultBrokerStrategy === "last" || isEmpty(settings?.defaultBrokerStrategy)) {
+		if (interopOverrideSettings?.defaultBrokerStrategy === "last") {
 			logger.info("Adding default interop override last");
-			allOverrides.push(await getConstructorOverride(settings));
+			allOverrides.push(await getConstructorOverride(interopOverrideSettings));
 		}
 		logger.info("Finished setting up interop overrides.");
 		isInitialized = true;
@@ -68,4 +85,3 @@ export function getInteropConstructorOverrides(): OpenFin.ConstructorOverride<Op
 	}
 	return allOverrides;
 }
-